requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
Django==3.0.8
django-environ==0.4.5
future==0.18.2
idna==2.10
line-bot-sdk==1.17.0
mysqlclient==2.0.1
PyMySQL==0.10.0
pytz==2020.1
requests==2.25.0
sqlparse==0.3.1
urllib3==1.26.2
